refactor(rss_builder): move Enclosure class to RssBuilder namespace
